use packaged pipeline spec in k8s platform pkg
this change uses the pipeline spec that is already included in the
kfp namespace, instead of expecting the pb2 generate pipeline spec to
be present in the kfp-kubernetes package. This avoids any conflicts
in the pipeline_spec pb2 symbols, since we are just using one source.

In order to enable intellisense behavior, the change also includes a pipeline_spec_pb2.py shim, this makes is so ide tools do not throw unresolved errors when trying to access pipeline spec symbols.

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>