Bump protobuf to 6.x for pipeline spec
This change bumps support for kfp-pipeline-spec to 6.x. This is a
significant bump by two major versions. We specifically opt not to support
a range (e.g. >=5.x,<=6.x) because we are bumping protoc to 30.x+ and this
compiler is only compatible with 6.x+.

Accompanying this change is a change to the makefile command. Previously,
the developer was required to hunt down and manually download the protos
originating from the protobuf libraries in git. The makefile will now
download these proto dependencies for the user.

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>