Bump protobuf to 6.x for kfp sdk
This change will bump KFP sdk to leverage protobuf 6.x. It also bumps
kfp-pipeline-spec to 0.8.0 which too requires protobuf 6.x.

Note that the version of KFP is also bumped to 2.4.0. This is required
because CI will fail otherwise when it attempts to retrieve KFP 2.3.0
due to a mismatch between kfp sdk 2.13 protobuf requirements (>=4.x,<5),
and kfp sdk 2.14 requirements (>=6.x, <7).

In order to accompany CI, we need to also ensure any CI that tries to
re-compile KFP from source must also install kfp-pipeline-spec from source,
otherwise it will try to download 0.8.0 before it is released. In order to
support this case, there is an env var introduced:

"KFP_PIPELINE_SPEC_PACKAGE_PATH", this var when set will have the executing
pipeline install pipeline-spec from the designated store instead of pypi
(transitively through KFP).

This change is also utilized in some of the kfp.local tests that will
compile and run tests locally via subprocess.

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>